requirements.txt
# ATTENTION!
# Use "pip install -r /path/to/requirements.txt" to install the dependencies.
# Use virtualenv for better results ;-)

django-tastypie==0.12.1

